fix: navbar docked, icon spacing, full dates, unit translations, polish
- Dock navbar flush to bottom
- Fix icon-to-text spacing in StatsTab
- Contest dates use full month names
- Replace hardcoded g/mL with i18n translations
- Add spaces between numbers and units
- Translate weight range pills
- Add top padding to ChatPage
- Remove redundant weight from Settings
- Fix Trends legend dot spacing

+++ src/pages/ContestPage.tsx
@@ -19,7 +19,7 @@ function formatDateRange(weekStart: string): string {
   end.setDate(end.getDate() + 6);
   const locale = i18n.language === "he" ? "he-IL" : "en-US";
   const fmt = (d: Date) =>
-    d.toLocaleDateString(locale, { month: "short", day: "numeric" });
+    d.toLocaleDateString(locale, { month: "long", day: "numeric" });
   return `${fmt(start)} - ${fmt(end)}`;
 }
